+++ plugins/pivora-core/src/blocks/slide-in-cta/index.js
@@ -12,6 +12,7 @@ import {
 } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
 import metadata from './block.json';
+import { DisplayConditionPanel } from '../shared/display-condition-panel';
 import './style.scss';
 
 function Edit( { attributes, setAttributes } ) {
@@ -78,6 +79,10 @@ function Edit( { attributes, setAttributes } ) {
 						}
 					/>
 				</PanelBody>
+				<DisplayConditionPanel
+					attributes={ attributes }
+					setAttributes={ setAttributes }
+				/>
 			</InspectorControls>
 			<div { ...blockProps } role="dialog" aria-modal="true">
 				<button